requirements.txt
pandas==2.0.0
requests==2.28.2
beautifulsoup4==4.12.2
selenium==4.8.3
pdfplumber==0.8.1
openpyxl==3.1.2
PyDrive==1.3.1
closeio~=2.0
oauth2client~=4.1.3
google-api-python-client~=2.86.0
Scrapy~=2.8.0
itemadapter~=0.8.0
undetected-chromedriver==3.4.6
cloudscraper~=1.2.71
cffi~=1.15.1
Pillow~=9.5.0
pytest~=7.3.1
lxml~=4.9.2
bs4~=0.0.1
soupsieve~=2.4.1
h11~=0.14.0
pip~=23.1.2
attrs~=23.1.0
wheel~=0.40.0
cryptography~=40.0.2
rsa~=4.9
pyasn1~=0.5.0
pytz~=2023.3
trio~=0.22.0
outcome~=1.2.0
sniffio~=1.3.0
exceptiongroup~=1.1.1
sortedcontainers~=2.4.0
idna~=3.4
Wand~=0.6.11
PyYAML~=6.0
numpy~=1.24.3
setuptools~=67.6.1
parse~=1.19.0
six~=1.16.0
cachetools~=5.3.0
certifi~=2022.12.7
urllib3~=1.26.15
python-dateutil~=2.8.2
scipy~=1.10.1
httplib2~=0.22.0
pyparsing~=3.0.9
hyperlink~=21.0.0
pycparser~=2.21
packaging~=23.1
zipp~=3.15.0
uritemplate~=4.1.1
wsproto~=1.2.0
pysat~=3.0.6
pyodbc~=4.0.39